+++ src/base/cs_setup.cpp
@@ -47,6 +47,7 @@
 #include "cs_atmo_variables.h"
 #include "cs_cf_thermo.h"
 #include "cs_coal_read_data.h"
+#include "cs_coupling.h"
 #include "cs_ctwr.h"
 #include "cs_ctwr_variables.h"
 #include "cs_domain_setup.h"
@@ -85,6 +86,7 @@
 #include "cs_rad_transfer_options.h"
 #include "cs_restart.h"
 #include "cs_runaway_check.h"
+#include "cs_syr_coupling.h"
 #include "cs_thermal_model.h"
 #include "cs_turbomachinery.h"
 #include "cs_turbulence_model.h"
@@ -2663,6 +2665,26 @@ _init_user
     if (cs_glob_atmo_option->soil_cat >= 0)
       cs_f_solcat(1);
   }
+
+  /* If thermal solver, set appropriate flags for coupling if needed,
+   * We do this before cs_user_parameters to allow the user to potentially
+   * modify this behavior if needed.
+   */
+  if (cs_glob_physical_model_flag[CS_HEAT_TRANSFER] > -1) {
+    if (cs_syr_coupling_n_couplings() > 0) {
+      /* We use minimal time step size while following the one
+       * "imposed" by the fluid solvers.
+       */
+      cs_log_printf(CS_LOG_DEFAULT,
+                    _("\n"
+                      " Setting Heat Transfer Solver to follow the time step\n"
+                      " value imposed by the fluid solver. This can be undone\n"
+                      " in cs_user_parameters if needed.\n"));
+      int _flag = PLE_COUPLING_TS_MIN | PLE_COUPLING_TS_FOLLOWER;
+      cs_coupling_set_sync_flag(_flag);
+    }
+  }
+
 
   /* User functions */
   cs_f_usipsu(nmodpp);